Fix sign compare in SimpleGraphics

+++ software/software_linux/libs/SimpleGraphics.cpp
@@ -26,8 +26,8 @@ void SimpleGraphics::draw_rect(rgba_t color, unsigned int x1, unsigned int y1, u
 #endif
 
 void SimpleGraphics::draw_x(rgba_t color, unsigned int x, unsigned int y, unsigned radius) {
-    for (int i = -(int) radius; i <= radius; i++) {
-        for (int j = -(int) radius; j <= radius; j++) {
+    for (int i = -(int) radius; i <= (int) radius; i++) {
+        for (int j = -(int) radius; j <= (int) radius; j++) {
             if (i == j || -i == j) {
                 draw_pixel(color, x + i, y + j);
             }
